+++ core/src/gzs/game/misc/MouseInfo.java
@@ -1,11 +1,11 @@
 package gzs.game.misc;
 
 public class MouseInfo {
-	private Pair<Double> lmp;
+	private Pair<Float> lmp;
 	private boolean md;
 	
-	public Pair<Double> getPosition() { return lmp; }
-	public void setPosition(double x, double y) {
+	public Pair<Float> getPosition() { return lmp; }
+	public void setPosition(float x, float y) {
 		this.lmp.x = x;
 		this.lmp.y = y;
 	}
@@ -14,7 +14,7 @@ public class MouseInfo {
 	public void setMouseDown(boolean md_) { this.md = md_; }
 	
 	public MouseInfo() {
-		lmp = new Pair<Double>(0.0, 0.0);
+		lmp = new Pair<Float>(0.0f, 0.0f);
 		md = false;
 	}
 }
