+++ core/src/gzs/game/objects/items/Item.java
@@ -9,14 +9,14 @@ import gzs.entities.Player;
 import gzs.game.misc.Pair;
 
 public abstract class Item implements Entity {
-	protected Pair<Double> position;
-	public Pair<Double> getPosition() { return position; }
+	protected Pair<Float> position;
+	public Pair<Float> getPosition() { return position; }
 	protected Texture icon;
 	
 	protected long duration;
 	protected long created;
 	
-	public Item(Pair<Double> pos, long cTime) {
+	public Item(Pair<Float> pos, long cTime) {
 		// If you do not use a custom constructor, the item will die instantly.
 		position = pos;
 		// You must also set the image or it will be null.
@@ -33,10 +33,10 @@ public abstract class Item implements Entity {
 	@Override
 	public void render(SpriteBatch batch, ShapeRenderer sr, long cTime) {
 		if(isActive(cTime) && (icon != null)) {
-			double x = position.x - (icon.getWidth() / 2);
-			double y = position.y - (icon.getHeight() / 2);
+			float x = position.x - (icon.getWidth() / 2);
+			float y = position.y - (icon.getHeight() / 2);
 			batch.begin();
-			batch.draw(icon, (float)x, (float)y);
+			batch.draw(icon, x, y);
 			batch.end();
 		}
 	}
@@ -48,8 +48,8 @@ public abstract class Item implements Entity {
 	
 	public boolean isTouching(double distance) {
 		if(icon != null) {
-			double width = icon.getWidth();
-			double height = icon.getHeight();
+			float width = icon.getWidth();
+			float height = icon.getHeight();
 			return (distance <= ((width + height) / 2));
 		} else return false;
 	}
