+++ core/src/gzs/game/objects/weapons/Weapon.java
@@ -20,7 +20,7 @@ public interface Weapon {
 	public abstract boolean hasWeapon();
 	public abstract void activate();
 	public abstract boolean canFire(long cTime);
-	public abstract void fire(Pair<Double> position, double theta, long cTime);
+	public abstract void fire(Pair<Float> position, float theta, long cTime);
 	public abstract void reload(long cTime);
 	public abstract boolean isReloading(long cTime);
 	public abstract double getReloadTime(long cTime);
